fix(placeholder): replace base64 placeholder with RGB dominant color
this change lighten the size of /api/books's payload from 683k to 4k
for 5 series with 6 images on average.
browsers are capable enough to create a canvas or an svg item
on the spot as a placeholder. my main focus right now is that
the application must be usable with slow internet.

the dominant color is calculated by create-db.

dependencies: remove gm, add color-thief

references:
https://futurestud.io/tutorials/how-to-get-dominant-color-code-for-picture-with-nodejs